Merge upstream changes from OBS 27.0.1 (#188)
* Fix macOS build error introduced in #236

This does not fix modifiers on macOS, that seems to be a bigger task.

* Fix ABI break introduced by adding panel Reload funtion

* CI: Update clang-format to support Ubuntu 18.04 and 20.04

* Clean up console message warnings

* Only use GPU SRGB if source texture supports it

* Include OBS version in agent string

This allows servers the ability to determine that browser panels and
browser sources are coming from OBS.

* Update browser version to 2.13.2

* Fix unnecessary pixel ratio calculations on macOS

MacOS reports correct window sizes with scaling applied already, no
further calculations are necessary. Changes introduced by 3cd1a98 are
also removed on macOS as the sub-window introduces another layer of
possible mismatches between CEF window and Qt window.

* CMake: Target macOS 10.13

* Add missing files check for local browser sources

* Update version to 2.13.3

* Show Reload button and hide Print from context menu

* Replace the CEF dropdown menu with a Qt dropdown

Items are copied from the CEF dropdown at runtime.

If new menu item types are used, the reference is available here:
https://magpcss.org/ceforum/apidocs3/projects/(default)/cef_menu_item_type_t.html

* Update version to 2.14.0

* Limit Qt browser panel context menu to Windows for now

On both macOS and Linux, the menu fails to close when clicking within
the browser panel (OnContextMenuDismissed doesn't trigger either).
As it requires further work, just disable it.

Additionally, this feature fails to build on Ubuntu 18.04 (but builds
fine on 20.04). Potentially a Qt version issue. More testing needed.

* Fix Shared Texture option in cmake-gui

When using set without specifying the CACHE argument, a normal/local
variable is set to the specified value. This results in the GUI showing
the CACHE variable value (OFF), but the rest of the CMake script checks
the normal variable value first. Thus, the script works as if
SHARED_TEXTURE_SUPPORT_ENABLED is set to ON, even though cmake-gui shows
it set to OFF.

* Fix USE_QT_LOOP option in cmake-gui

See parent commit for full explanation.

* Translate Reload/Refresh context menu item

* Fix CEF locale path on *nix platforms

This fixes an issue where the browser panel context menu would
have blank text, making it near-impossible to use.
This issue does not occur on Windows. macOS uses a special path
for Framework locales, and handles *most* languages without issue.

This solution does not work if the locale path is relative.

* Improve rendering performance when not using shared textures

Fixes #278

This flag was used before the Browser 2.0 rewrite, and never revisited.
After testing across all platforms, removing this flag allows access to
OpenGL even when shared textures are unavailable. At the moment, this is
especially beneficial for Linux users.

* Disable browser panels on Wayland

Since the browser panels currently don't work on Wayland,
disable them for now, if the user is running Wayland.

* Emit fatal error if CEF or X11 is missing

CMake will now emit a fatal error if CEF or X11
is not found rather than disabling obs-browser

* Fix bitmask typo when creating OBS version string

Detected by PVS Studio.

* Fix memory corruption when translating Refresh text

* Update version to 2.14.2

* Add SRGB support flag

This flag indicates to OBS that the source is SRGB-aware, and not to
force-disable it.

* Update version to 2.14.3

* Remove clang-format github workflow

* Build against obs-studio 27.0.1

* Windows build with dependencies2019.zip

Co-authored-by: Matt Gajownik <matt@obsproject.com>
Co-authored-by: jpark37 <jpark37@users.noreply.github.com>
Co-authored-by: jp9000 <obs.jim@gmail.com>
Co-authored-by: PatTheMav <PatTheMav@users.noreply.github.com>
Co-authored-by: Ryan Foster <RytoEX@gmail.com>
Co-authored-by: Matt Gajownik <matt@wizardcm.com>
Co-authored-by: Clayton Groeneveld <claytong1214@gmail.com>
Co-authored-by: tytan652 <tytan652@tytanium.xyz>
Co-authored-by: Richard Stanway <r.stanway@gmail.com>