Downgrade cookies from CEF 4638 on Windows (#195)
In case we failed to decrypt some cookie rows (there is an issue with
this in recent versions of CEF), delete cookie rows with encrypted
values before the cookies file has a chance to be referenced by a
`CefCookieManager`.

We must do this, since the version of CEF we bundle with the product at
this time differs from the version bundled with OBS Studio 27.2 (4638).

Any version of CEF, starting with 3770 and above, stores cookies in
an encrypted column.

CEF 3729, which we bundle with SE.Live does not support encrypted
cookies, so until this moment we used Win32 DPAPI to decrypt cookies
from CEF 3770, and store them in plain text for compatibility.

CEF 4638 bundled with OBS Studio 27.2 on Windows no longer supports
decryption with DPAPI.