Simplify how dock widgets are created and visibility is set
We no longer do `QApplication::sendPostedEvents()` when setting the visibility of a dock widget.
Instead, we moved `widget->setVisible(...)` into `QtPostTask()` to make sure it is
executed after the event loop has run.

In addition, we removed a redundant call to `QMainWindow::setCentralWidget(widget)` which
was called by `StreamElementsBrowserWidgetManager::AddDockBrowserWidget`. Most likely this
is the remains of a copy & paste error which took place years ago.

All those should have the happy side effect of creating `floating` dock widgets initially
visible _in front_ of the OBS main window and not behind it until the user clicks on the
OBS main window.