Make audio in custom canvases work
This requires building an audio root transition wrapper source, which only
extracts audio from the transition, and assigning it to an available
output channel.

We have to do this, since if we assign the root transition to the output
channel, it will overlay the video in the main, native OBS canvas.