Safer properties retrieval when enumerating existing sources
Some sources do not support retrieval of properties by
source id (type), since in that case, the argument which is
passed to the get_properties source function is NULL.

Sources like Spout and Bongobs Cat Plugin do not support this
way of retrieving properties, and throw an exception.

To mitigate this, we only retireve properties for a fully-instantiated
source.