Make restored dock widgets geometry impact on main OBS window size less annoying
We do this by comparing the OBS main window height with it's screen's available height,
and if it's larger than the available screen height, we adjust the main window height
to fit the available screen height.

In this case we also make sure to position the main window at the top of the screen.