Refactor scenes and scene items signal handling code (#36)
Previously, not all signal handlers had a reference to `SESignalHandler`
context, which holds the video composition, scene manager and relevant
root scene context.

This resulted in limitations and sometimes unpredictable behavior when
processing signals for scene items which are located under Groups, which
are implemented as Scenes themselves internally by OBS.

One of the limitations was that there is no way to resolve a root parent
scene from a scene item which is located under a Group. This caused
all sorts of inconsistencies in JS API event payloads, and sometimes
led to crashes, supposedly due to inconsistencies in OBS internal state.

In addition, there was no way of tracking asynchronous operations
completion - i.e. when asynchronous signal handling code, built to
circumvent deadlocks due to OBS API internal locking shenanigans
has completed running.

This update solves all those issues: all signal handlers now have the
proper root scene context, asynchronous signal handling is tracked and
waited for, and references to `SESignalHandler` are tracked and
accounted for, to make sure that we are not leaking signal handler
registrations.

https://app.bugsplat.com/v2/crash?database=OBS_Live&id=1490737