API 6.1: Added event at the beginning of scene change sequence (#37)
Add Events:

- `hostActiveSceneChanging`

The only event reported by the OBS front-end which is related to
scene changes, is fired when scene change has completed - i.e.
the transition sequence from scene A to scene B has completed.

A transition can take a long time to complete, based on user
settings. This means that if we want to react to the users'
_intention_ to change scenes instead of a completed transition
sequence, we need a different mechanism.

We implement it by listening to the current OBS transition's
`transition_start` signal, and based on whether we are in OBS
studio mode or not, either query the OBS frontend API for the
current selected Scene (we have to do this because in studio
mode, transition target scenes do not match the ones used in
normal mode), _or_, in non-studio mode, query the current OBS
transition for it's second (target) source.

There still may be edge cases when the transition bar is used
in OBS studio mode, however, they require a deeper product
thought to tackle and are left for a later time.