Stop using deprecated global OBS frontend settings
`obs_frontend_get_global_config()` was deprecated in favor
of `obs_frontend_get_user_config()` in OBS 31.

There is no graceful transition path: all user settings have been
moved to `obs_frontend_get_user_config()`, so we had to upgrade to
building against OBS 31, and replace the function used to retrieve
user settings simultaneously.

This is not ideal. We usually like to wait until an OBS major release
stabilizes and the next major version update is released, but it
is what it is.