SE.Live API WebSocket server will now reply to the client which made a call
This will remove inconsistencies in callback IDs which are an ever-incrementing
integer number, specific to each WebSocket client.

This should mitigate possible inconsistencies between different clients with
the same client ID, and different unique client IDs.

An example of such case is a browser widget and a chromium debugger inspector
attached to that browser.