Replay Buffer: dependency on Streaming state and default minimum file time (#45)
When a replay buffer is active, it reflects on the OBS frontend UI: i.e.
recording settings cannot be updated, while at least one replay buffer is active.
For this reason, we make sure our replay buffers can only be active when
OBS is streaming.

In addition, if replay buffer duration is zero or less than a keyframe interval,
attempt to save it to a file will result in OBS getting stuck in an infinite
loop trying to purge excess packets from the buffer forever.
If we detect that replay buffer duration is 0 or not specified, we'll set it
to 20 seconds.